fix: Better name for query param
